refactor: codeview moudlarizing
